renamed calc_averages
